+++ src/main/java/name/richardson/james/bukkit/utilities/command/argument/SimpleArgumentMetadata.java
@@ -18,12 +18,9 @@
 
 package name.richardson.james.bukkit.utilities.command.argument;
 
-import name.richardson.james.bukkit.utilities.localisation.Localisation;
-import name.richardson.james.bukkit.utilities.localisation.StrictResourceBundleLocalisation;
+import name.richardson.james.bukkit.utilities.localisation.Localised;
 
 public class SimpleArgumentMetadata implements ArgumentMetadata {
-
-	private final Localisation localisation = new StrictResourceBundleLocalisation();
 
 	private final String id;
 	private final String name;
@@ -31,11 +28,17 @@ public class SimpleArgumentMetadata implements ArgumentMetadata {
 
 	private String error;
 
-	public SimpleArgumentMetadata(final String id, final String name, final String desc, final String error) {
-		this.id = localisation.getMessage(id);
-		this.name = localisation.getMessage(name);
-		this.desc = localisation.getMessage(desc);
-		if (error != null) this.error = localisation.getMessage(error);
+	public SimpleArgumentMetadata(Localised id, Localised name, Localised desc, Localised error) {
+		this.id = id.asMessage();
+		this.name = name.asMessage();
+		this.desc = desc.asMessage();
+		this.error = error.asMessage();
+	}
+
+	public SimpleArgumentMetadata(Localised id, Localised name, Localised desc) {
+		this.id = id.asMessage();
+		this.name = name.asMessage();
+		this.desc = desc.asMessage();
 	}
 
 	@Override