+++ src/main/java/name/richardson/james/bukkit/utilities/formatters/AbstractChoiceFormatter.java
@@ -6,9 +6,6 @@ import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 
-import name.richardson.james.bukkit.utilities.localisation.Localisation;
-import name.richardson.james.bukkit.utilities.localisation.StrictResourceBundleLocalisation;
-
 public abstract class AbstractChoiceFormatter implements ChoiceFormatter {
 
 	private final Localisation localisation = new StrictResourceBundleLocalisation();