+++ src/main/java/name/richardson/james/bukkit/utilities/persistence/database/AbstractDatabaseMigrator.java
@@ -21,9 +21,7 @@ package name.richardson.james.bukkit.utilities.persistence.database;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import name.richardson.james.bukkit.utilities.localisation.Localisation;
 import name.richardson.james.bukkit.utilities.localisation.PluginLocalisation;
-import name.richardson.james.bukkit.utilities.localisation.StrictResourceBundleLocalisation;
 import name.richardson.james.bukkit.utilities.logging.PluginLoggerFactory;
 
 /**